MODIFY: rename resource to controllers, add handlers for voting and querying all cuisine
